+++ lib/src/gg_localize_refs.dart
@@ -5,8 +5,8 @@
 // found in the LICENSE file in the root of this package.
 
 import 'package:args/command_runner.dart';
-import 'package:gg_localize_refs/src/commands/unlocalize_refs.dart';
-import 'commands/localize_refs.dart';
+import 'package:gg_localize_refs/src/commands/change_refs_to_pub_dev.dart';
+import 'package:gg_localize_refs/src/commands/change_refs_to_local.dart';
 import 'package:gg_log/gg_log.dart';
 import 'package:gg_localize_refs/src/commands/get_ref_version.dart';
 import 'package:gg_localize_refs/src/commands/set_ref_version.dart';
@@ -16,8 +16,8 @@ import 'package:gg_localize_refs/src/commands/get_version.dart';
 class GgToLocal extends Command<dynamic> {
   /// Constructor
   GgToLocal({required this.ggLog}) {
-    addSubcommand(LocalizeRefs(ggLog: ggLog));
-    addSubcommand(UnlocalizeRefs(ggLog: ggLog));
+    addSubcommand(ChangeRefsToLocal(ggLog: ggLog));
+    addSubcommand(ChangeRefsToPubDev(ggLog: ggLog));
     addSubcommand(GetRefVersion(ggLog: ggLog));
     addSubcommand(SetRefVersion(ggLog: ggLog));
     addSubcommand(GetVersion(ggLog: ggLog));